FIX (z9999): tree density setting form uint8 to uint16
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@1276 8aca7d54-2c30-db11-9de9-000461428c89

+++ dings/baum.cc
@@ -56,7 +56,7 @@ uint8 baum_t::forest_boundary_blur = 6;
 uint8 baum_t::forest_boundary_thickness = 2;
 
 // Determins how often are spare trees going to be planted (works inversly)
-uint8 baum_t::forest_inverse_spare_tree_density = 5;
+uint16 baum_t::forest_inverse_spare_tree_density = 5;
 
 // Number of trees on square 2 - minimal usable, 3 good, 5 very nice looking
 uint8 baum_t::max_no_of_trees_on_square = 3;
